Replace HttpAssetAccessor with UrlAssetAccessor
The motivation for this change is to take advantage of reusing the
CURL handles, rather than tearing them down and creating them anew for
each request. This allows libcurl to keep a connection to a remote
server open and reuse it for multiple requests. This avoids
renegotating SSL among other things and can be a big performance win.